pyproject.toml
[project]
name = "spicy"
version = "0.1.0"
description = "Almost an mdneeds"
readme = "README.md"
authors = [
    { name = "Richard Fabian", email = "raspo1@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "markdown-it-py>=3.0.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-forked>=1.6.0",
    "pytest-xdist>=3.6.1",
]

[project.scripts]
spicy = "spicy:main"
spicecheck = "spicy.check_use_cases:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
